rabbitmq-run.mk: Re-use `test-dist` if the build was already done for tests
The testsuites default to run `make test-dist` as a first step.
Therefore later, when it starts a node, it should re-use that isntead of
depending on `make dist` which will rebuild the tested project and
remove test dependencies from plugins.

This is also useful for the rabbitmq-server(8) replacement effort
because, in the case of rabbitmq-erlang-client, the broker is a
$(TEST_DEPS): if starting a node runs `make dist`, the broker will be
removed. However, in this new world, it is expected to be packaged as an
.ez archive.